refactor(core): resolve all compilers warnings

+++ repl.c
@@ -9,7 +9,7 @@
  */
 void tsh_repl(tsh_t *tsh)
 {
-  while (read_command(tsh) != -1)
+  while (read_command(tsh) != (size_t) -1)
   {
     command_t *command;
